Move calculate winner logic to reusable component
