+++ src/components/App.js
@@ -4,11 +4,7 @@ import Game from "./Game";
 import "../index.css";
 
 const App = () => {
-  return (
-    <div>
-      <Game />
-    </div>
-  );
+  return <Game />;
 };
 
 export default App;
