+++ src/components/Game.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 
 import Board from "./Board";
 import RulesModal from "./RulesModal";
+import calculateWinner from "../Help";
 
 const Game = () => {
   const [stepNumber, setStepNumber] = useState(0);
@@ -96,7 +97,6 @@ const Game = () => {
               />
             </div>
           </div>
-
           <p>{statusText}</p>
           <Board
             squares={current.squares}
@@ -113,23 +113,3 @@ const Game = () => {
 };
 
 export default Game;
-
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
