+++ src/components/Square.js
@@ -1,9 +1,9 @@
 import React from "react";
 
-const Square = props => {
+const Square = ({ onClick, value }) => {
   return (
-    <button className="square" onClick={props.onClick}>
-      {props.value}
+    <button className="square" onClick={onClick}>
+      {value}
     </button>
   );
 };
